requirements.txt
python-telegram-bot==20.7
requests
beautifulsoup4
python-dotenv
selenium
python-dateutil
colorlog
meta-ai-api
python-telegram-bot[job-queue]